refactor(decomp): extract cdc connectors → proximadb-cdc (TD-DECOMP-12)
Moves src/cdc/connectors/{mongodb,mysql,postgres}/* (93 tests) into the cdc
crate (199→292 tests). Whole subdir; rewrite crate::cdc::→crate:: (all
refs). Added: experimental-cdc-connectors feature (dep:tokio-postgres),
anyhow (workspace), base64 (inline 0.22), tokio-postgres (inline 0.7
optional). DB-driver code (mongodb/bson/mysql_async) is behind cfg(feature
= "experimental-cdc-connectors") — never enabled in CI, so default builds
compile without those heavy deps.

Verified: cdc crate 292 tests pass; root check + server build + cdc-clippy
+ fmt + check_workspace_boundaries + check-layering all green.
