+++ crates/integrations/proximadb-cdc/Cargo.toml
@@ -16,6 +16,7 @@ path = "src/lib.rs"
 [features]
 # Kafka sink uses rdkafka when enabled; stubs otherwise (mirrors root's cdc-kafka feature).
 cdc-kafka = []
+experimental-cdc-connectors = ["dep:tokio-postgres"]
 
 [dependencies]
 # `event` uses serde (derive) + serde_json (Value/ser) + rand (event-id gen).
@@ -26,6 +27,8 @@ serde.workspace = true
 serde_json.workspace = true
 tokio.workspace = true
 proximadb-config.workspace = true
+anyhow.workspace = true
+base64 = "0.22"
 # sinks: async-trait (CdcSink trait), tracing (logging), reqwest (webhook HTTP).
 async-trait.workspace = true
 tracing.workspace = true
@@ -33,6 +36,8 @@ reqwest = { version = "0.11", features = ["json"] }
 # outbound: uuid (position/event IDs), regex (event filtering).
 uuid = { version = "1.0", features = ["v4"] }
 regex.workspace = true
+# connectors (experimental): tokio-postgres behind the experimental-cdc-connectors feature.
+tokio-postgres = { version = "0.7", optional = true }
 
 
 [dev-dependencies]
