+++ crates/integrations/proximadb-cdc/src/connectors/mysql/config.rs
