+++ docs/10-quality/td/TD-DECOMP-12-extract-cdc-connectors.adoc
@@ -1,0 +1,37 @@
+// Copyright (C) 2025 ProximaDB
+// SPDX-License-Identifier: Apache-2.0
+= TD-DECOMP-12: Extract cdc connectors → proximadb-cdc (god-crate decomposition)
+
+[cols="1,3", options="header"]
+|===
+| Field | Value
+| Status | Landing
+| Severity | Low — behavior-neutral re-export
+| Component | `crates/integrations/proximadb-cdc` (`connectors`); root `src/cdc/mod.rs`
+| Relates to | [[TD-DECOMP-1]] (recipe); [[TD-DECOMP-7..11]] (cdc crate prior slices)
+|===
+
+== Why
+
+The connectors subgraph (`{mongodb,mysql,postgres}/*`, **93 tests**) is the last big cdc
+subgraph. It references `crate::cdc::{config,error,event,offset,source}` (all now in the cdc
+crate via TD-DECOMP-7..11). DB-driver code (mongodb/bson/tokio_postgres/mysql_async) is behind
+`#[cfg(feature = "experimental-cdc-connectors")]` (never enabled in CI) so the default build
+compiles without those heavy deps.
+
+== What moved (behavior-neutral)
+
+* `src/cdc/connectors/` (whole subdir: mod + mongodb/{mod,change_event,config,connector} +
+  mysql/{mod,config,connector,decoder} + postgres/{mod,config,connector,decoder,replication,
+  snapshot}) → `crates/integrations/proximadb-cdc/src/connectors/`.
+* Rewrite: `crate::cdc::` → `crate::` (all refs — config/error/event/offset/source become
+  siblings; connectors self-refs stay valid).
+* cdc crate: + `experimental-cdc-connectors` feature (`["dep:tokio-postgres"]`), + deps
+  `anyhow` (workspace), `base64` (inline `"0.22"`), `tokio-postgres` (inline `{ version =
+  "0.7", optional = true }`).
+* cdc crate 199→292 tests.
+
+== Follow-ups
+
+* `coordinator` (~11 tests, the hub) — the LAST cdc module; completes the cdc extraction.
+* **Running total once merged: ~417 tests out of root lib binary** (324 + 93).
